If this is it...
Then I'm fucking stupid and need to stop making website. If this is really it then it took me too long to fucking realize that that's what it meant. I kept thinking it either wanted an any type or something else like that... yeah idk why so don't ask me why.